moved the tests to their correct locations
